worker/requirements.txt
psycopg2-binary==2.9.3
pika==1.3.0
python-dotenv==0.20.0